pyproject.toml
[project]
name = "final-tax-return"
version = "1.0.0"
description = "確定申告用収支記録アプリケーション"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dash>=2.14.0",
    "dash-bootstrap-components>=1.5.0",
    "pandas>=2.0.0",
    "plotly>=5.18.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
    "pypdf2>=3.0.0",
    "supabase>=2.0.0",
    "gunicorn>=21.0.0",
]

[project.scripts]
tax-app = "app:main"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pandas-stubs>=2.0.0",
]

# =============================================================================
# Ruff Configuration
# =============================================================================
[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "SIM108", # Use ternary operator (sometimes less readable)
]

[tool.ruff.lint.isort]
known-first-party = ["app", "utils", "pages", "components"]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"

# =============================================================================
# Mypy Configuration
# =============================================================================
[tool.mypy]
python_version = "3.10"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
check_untyped_defs = true
ignore_missing_imports = true

# Third-party libraries without stubs
[[tool.mypy.overrides]]
module = [
    "dash.*",
    "dash_bootstrap_components.*",
    "dash_table.*",
    "plotly.*",
    "google.*",
    "google.genai.*",
    "supabase.*",
    "PIL.*",
    "PyPDF2.*",
]
ignore_missing_imports = true

# Files with complex Dash/Gemini interactions (type stubs incomplete)
[[tool.mypy.overrides]]
module = [
    "utils.gemini",
    "pages.records.layout",
    "pages.records.callbacks",
    "index",
]
ignore_errors = true
